Move interface into folder
