+++ sample/000-dogs-app/src/dog/interfaces/dog.interface.ts
